requirements.txt
flask==2.0.1
opencv-python==4.5.3.56
pyzbar==0.1.8
gspread==5.0.0
oauth2client==4.1.3
qrcode==7.3
pillow==8.3.2
pandas==1.3.3
